polishing code and adding password
