+++ export_manager.py
@@ -6,6 +6,7 @@ import shutil
 
 from bpy_extras.io_utils import ExportHelper
 from bpy.types import Operator
+from bpy.types import Panel, UIList
 
 from bpy.props import EnumProperty, StringProperty, BoolProperty, IntProperty, CollectionProperty, BoolVectorProperty, PointerProperty
 import bpy.props as prop
@@ -13,6 +14,118 @@ import bpy.props as prop
 from .functions import *
 
 import random
+
+
+#####################################################################
+#Export Section
+
+class EM_ExportPanel:
+    bl_label = "Export Manager"
+    bl_space_type = 'VIEW_3D'
+    bl_region_type = 'UI'
+
+    def draw(self, context):
+        layout = self.layout
+        scene = context.scene
+ 
+
+        row = layout.row()
+        box = layout.box()
+        row = box.row()
+        row.label(text="Tables export:")
+        row = box.row()
+        op = row.operator("export.uuss_export", text="EM (csv)", emboss=True, icon='LONGDISPLAY')
+        row = box.row()
+        row.prop(context.window_manager.export_tables_vars, 'table_type', expand=True)
+        row = layout.row()
+        box = layout.box()
+        row = box.row()
+        row.label(text="EMviq export:")
+        row = box.row()
+        row.prop(context.scene, 'EMviq_project_name', toggle=True, text="")
+        row.label(text="<-- Project's name")
+        row = box.row()#(align=True)
+        row.prop(context.scene, 'EMviq_model_author_name', toggle=True, text="")
+        row.label(text="<-- Model's author(s)")
+        row = box.row()
+        row.prop(context.scene, 'EMviq_user_name', toggle=True, text="")
+        row.label(text="<-- ATON user's name")
+        row = box.row()
+        row.prop(context.scene, 'password', toggle=True, text="")
+        row.label(text="<-- ATON user's password")
+        row = box.row()
+        row.prop(context.scene, 'ATON_path', toggle=True, text="")
+        row.label(text="<-- path to ATON")
+
+        '''
+        row = layout.row()  # (align=True)
+        row.prop(context.scene, 'EMviq_folder', toggle=True, text="")
+        row.label(text="<-- Collection folder export path")
+        row = layout.row()#(align=True)
+        row.prop(context.scene, 'EMviq_scene_folder', toggle=True, text="")
+        row.label(text="<-- Scene folder export path")
+        '''
+
+        row = box.row()
+        row.prop(context.window_manager.export_vars, 'format_file', expand=True)
+        #box = layout.box()
+        row = box.row()
+        row.prop(context.scene, 'enable_image_compression', toggle = True, text='Use tex compression')
+        row.prop(context.scene, 'EM_gltf_export_maxres', toggle = True, text='Max res')
+        row.prop(context.scene, 'EM_gltf_export_quality', toggle = True, text='Size')
+        row = box.row()
+        op = row.operator("export_manager.export", text="Generate full EMviq Project", emboss=True, icon='LONGDISPLAY')
+        op.em_export_type = 'EMviq'
+        op.em_export_format = context.window_manager.export_vars.format_file
+
+        row = box.row()
+        row.label(text="Export partial Emviq project's components:")
+        row = box.row()
+        op = row.operator("export_manager.export", text="Proxies", emboss=True, icon='SHADING_SOLID')
+        op.em_export_type = 'Proxies'
+        op = row.operator("export_manager.export", text="GraphML", emboss=True, icon='SHADING_SOLID')
+        op.em_export_type = 'GraphML'
+        op = row.operator("export_manager.export", text="RM", emboss=True, icon='SHADING_TEXTURE')
+        op.em_export_type = 'RM'
+
+        row = layout.row()
+
+        if scene.emviq_error_list_index >= 0 and len(scene.emviq_error_list) > 0:
+            row.template_list("ER_UL_List", "EM nodes", scene, "emviq_error_list", scene, "emviq_error_list_index")
+            item = scene.emviq_error_list[scene.emviq_error_list_index]
+            box = layout.box()
+            row = box.row(align=True)
+
+            split = row.split()
+            col = split.column()
+            row.prop(item, "name", text="")
+            split = row.split()
+            col = split.column()
+            op = col.operator("select.fromlistitem", text='', icon="MESH_CUBE")
+            op.list_type = "emviq_error_list"
+            #op = col.operator("listitem.toobj", icon="PASTEDOWN", text='')
+            #op.list_type = "emviq_error_list"
+            #row = layout.row()
+            #row.label(text="Description:")
+            row = box.row()
+            #layout.alignment = 'LEFT'
+            row.prop(item, "description", text="", slider=True, emboss=True)
+
+class VIEW3D_PT_ExportPanel(Panel, EM_ExportPanel):
+    bl_category = "EM"
+    bl_idname = "VIEW3D_PT_ExportPanel"
+    bl_context = "objectmode"
+
+class ER_UL_List(bpy.types.UIList):
+    def draw_item(self, context, layout, data, item, icon, active_data, active_propname, index):
+        icons_style = 'OUTLINER'
+        scene = context.scene
+        layout = layout.split(factor =0.30, align = True)
+        layout.label(text = item.texture_type, icon = 'ERROR')
+        layout.label(text = item.material, icon='NONE', icon_value=0)
+
+#Export Section
+#####################################################################
 
 def check_if_scalable(image_block):
     is_scalable = False
@@ -445,8 +558,6 @@ def convert_shape2type(yedtype):
         node_type = ["USD","Documentary Stratigraphic Unit"]
     return nodetype
 
-
-
 def write_UUSS_data(context, filepath, only_UUSS, header):
     print("running write some data...")
     
@@ -560,6 +671,8 @@ def createfolder(base_dir, foldername):
     return export_folder
 
 classes = [
+    ER_UL_List,
+    VIEW3D_PT_ExportPanel,
     EM_export,
     ExportuussData,
     OBJECT_OT_ExportUUSS,
@@ -573,7 +686,7 @@ def register():
     bpy.types.Scene.password = bpy.props.StringProperty(
         name = "password Aton user",
         default = "",
-        description = "Set here your password to access an Aton instance"
+        description = "Set here your password to access an Aton instance",
         subtype='PASSWORD'
     )
